+++ tools/__init__.py
@@ -1,9 +1,9 @@
-from tools.rag_tool import rag_retrieve, grade_document, rewrite_query
+from tools.rag_tool import rag_retrieve, grade_documents, rewrite_query
 from tools.web_search_tool import web_search
 
 __all__ = [
     "rag_retrieve",
-    "grade_document",
+    "grade_documents",
     "rewrite_query",
     "web_search",
 ]
